+++ django_openS3/storage.py
@@ -1,10 +1,15 @@
 """
 A custom Storage interface for storing files to S3 via OpenS3
 """
+import datetime
 from django.conf import settings
+from django.core.exceptions import ImproperlyConfigured
 from django.core.files.base import ContentFile
 from django.core.files.storage import Storage
-import openS3
+from openS3 import OpenS3
+from openS3.constants import AWS_DATETIME_FORMAT
+from openS3.utils import S3IOError
+from django_openS3.constants import S3_STATIC_DIR, S3_MEDIA_DIR
 
 
 class S3Storage(Storage):
@@ -14,13 +19,22 @@ class S3Storage(Storage):
     up to or down from AWS.
     """
     def __init__(self, name_prefix, bucket, aws_access_key, aws_secret_key):
-        self.name_prefix = name_prefix
-        self.opener = openS3.OpenS3(bucket, aws_access_key, aws_secret_key)
+        self.name_prefix = None
+        if name_prefix:
+            if len(name_prefix) < 3:
+                raise ImproperlyConfigured('Given name prefix is too short. '
+                                           'Given name prefix {}'.format(name_prefix))
+            if name_prefix[0] != '/' or name_prefix[-1] != '/':
+                raise ImproperlyConfigured('Given name prefix must start and end with a slash. '
+                                           'Given name prefix {}'.format(name_prefix))
+            self.name_prefix = name_prefix
+        self.opener = OpenS3(bucket, aws_access_key, aws_secret_key)
 
     def _open(self, name, mode='rb'):
         """
         Retrieves the specified file from storage.
         """
+        name = self._prepend_name_prefix(name)
         with self.opener(name, mode) as fd:
             content = fd.read()
         return ContentFile(content, name)
@@ -30,79 +44,83 @@ class S3Storage(Storage):
         name = self._prepend_name_prefix(name)
 
         with self.opener(name, 'wb') as fd:
-            fd.write(content)
+            fd.write(content.read().decode())
+        return name
+
+    def get_valid_name(self, name):
+        # TODO Implement this method in a AWS friendly way.
+        # Don't use Django's get_valid_name as it strips slashes from the name.
         return name
 
     def _prepend_name_prefix(self, name):
         """Return file name (ie. path) with the prefix directory prepended"""
         if not self.name_prefix:
             return name
-        base = self.name_prefix.strip('/')
-        base = '/' + base
-        if name[0] != '/':
-            name = '/' + name
+        base = self.name_prefix
+        if name[0] == '/':
+            name = name[1:]
         return base + name
-
-    # The following methods form the public API for storage systems, but with
-    # no default implementations. Subclasses must implement *all* of these.
 
     def delete(self, name):
         """
         Deletes the specified file from the storage system.
         """
-        raise NotImplementedError('subclasses of Storage must provide a delete() method')
+        name = self._prepend_name_prefix(name)
+        with self.opener(name) as fd:
+            fd.delete()
 
     def exists(self, name):
         """
         Returns True if a file referenced by the given name already exists in the
         storage system, or False if the name is available for a new file.
         """
-        raise NotImplementedError('subclasses of Storage must provide an exists() method')
+        name = self._prepend_name_prefix(name)
+        with self.opener(name) as fd:
+            return fd.exists()
 
     def listdir(self, path):
         """
         Lists the contents of the specified path, returning a 2-tuple of lists;
         the first item being directories, the second item being files.
         """
-        raise NotImplementedError('subclasses of Storage must provide a listdir() method')
+        with self.opener(path) as fd:
+            return fd.listdir()
 
     def size(self, name):
         """
         Returns the total size, in bytes, of the file specified by name.
         """
-        raise NotImplementedError('subclasses of Storage must provide a size() method')
+        name = self._prepend_name_prefix(name)
+        with self.opener(name) as fd:
+            return fd.size
 
     def url(self, name):
         """
         Returns an absolute URL where the file's contents can be accessed
         directly by a Web browser.
         """
-        raise NotImplementedError('subclasses of Storage must provide a url() method')
-
-    def accessed_time(self, name):
-        """
-        Returns the last accessed time (as datetime object) of the file
-        specified by name.
-        """
-        raise NotImplementedError('subclasses of Storage must provide an accessed_time() method')
-
-    def created_time(self, name):
-        """
-        Returns the creation time (as datetime object) of the file
-        specified by name.
-        """
-        raise NotImplementedError('subclasses of Storage must provide a created_time() method')
+        name = self._prepend_name_prefix(name)
+        return self.opener(name).url
 
     def modified_time(self, name):
         """
         Returns the last modified time (as datetime object) of the file
         specified by name.
         """
-        raise NotImplementedError('subclasses of Storage must provide a modified_time() method')
+        name = self._prepend_name_prefix(name)
+        with self.opener(name) as fd:
+            fd.exists()  # Call .exists() to populate .response_headers
+            response_headers = fd.response_headers
+
+        if 'Last-Modified' not in response_headers:
+            raise S3IOError('"Last-Modified" header not found for S3 object {}'.format(name))
+        datetime_str = response_headers['Last-Modified']
+        return datetime.datetime.strptime(datetime_str, AWS_DATETIME_FORMAT)
 
 
 class S3StaticStorage(S3Storage):
-    def __init__(self, name_prefix=settings.STATIC_DIR,
+    def __init__(self,
+                 name_prefix=getattr(settings, 'S3_STATIC_DIR', None) or S3_STATIC_DIR,
                  bucket=settings.AWS_STORAGE_BUCKET_NAME,
                  aws_access_key=settings.AWS_ACCESS_KEY_ID,
                  aws_secret_key=settings.AWS_SECRET_ACCESS_KEY):
@@ -110,7 +128,8 @@ class S3StaticStorage(S3Storage):
 
 
 class S3MediaStorage(S3Storage):
-    def __init__(self, name_prefix=settings.MEDIA_DIR,
+    def __init__(self,
+                 name_prefix=getattr(settings, 'S3_MEDIA_DIR', None) or S3_MEDIA_DIR,
                  bucket=settings.AWS_STORAGE_BUCKET_NAME,
                  aws_access_key=settings.AWS_ACCESS_KEY_ID,
                  aws_secret_key=settings.AWS_SECRET_ACCESS_KEY):